+++ app.rb
@@ -4,56 +4,9 @@ require "lazyrecord"
 require "http"
 require_relative "model/books"
 require_relative 'model/googleapi'
-require_relative 'helpers/noimage'
-require_relative 'helpers/status'
-require_relative 'model/books'
-helpers NoNil, Status
+require_relative 'helpers/helpers_utils'
 
-class GoogleData
-  attr_accessor :data
-  def initialize(id_google)
-    @id_google = id_google
-  end
-
-  def load_data
-    @data = HTTP.get("https://www.googleapis.com/books/v1/volumes/#{@id_google}").parse
-  end
-
-  def title
-    @data["volumeInfo"]["title"]
-  end
-
-  def image
-    @data["volumeInfo"]["imageLinks"]["smallThumbnail"]
-  end
-
-  def subtitle
-    @data["volumeInfo"]["subtitle"]
-  end
-
-  def description
-    @data["volumeInfo"]["description"]
-  end
-
-  def author
-    @data["volumeInfo"]["authors"].join
-  end
-
-  def price
-    @data["saleInfo"]["listPrice"]["amount"].to_s << " " << @data["saleInfo"]["listPrice"]["currencyCode"]
-  end
-
-  def play_store
-    @data["volumeInfo"]["infoLink"]
-  end
-end
-
-helpers do
-  def html_option(value, label, current_value)
-    selected = value == current_value
-    "<option value=\"#{value}\" #{"selected" if selected}>#{label}</option>"
-  end
-end
+helpers NoNil, Status, HtmlHelper
 
 
 get "/books/:book_id" do
@@ -80,15 +33,15 @@ post "/books/:book_id/edit" do
 end
 
 get "/search" do
-  book=params["searchbook"].split(' ').join('+') rescue  nil
+  book = params["searchbook"].split(' ').join('+') rescue nil
   params["page"] ? count = 2 : count = 1
-  book && book != "" ? books = BooksApi::search_book(book,count) : books = nil
-  erb :search, locals: { books: books, search: book, page: count }
+  book && book != "" ? books = BooksApi::search_book(book, count) : books = nil
+  erb :search, locals: {books: books, search: book, page: count}
 end
 
 post "/book/register/" do
   id_goolge = params["id_google"]
   status = params["status"]
-  Book.new(id_goolge,status).save
+  Book.new(id_goolge, status).save
   redirect url("/search")
 end
