--- helpers/noimage.rb
@@ -1,8 +1,0 @@
-module NoNil
-  def image? (book)
-    book["volumeInfo"]["imageLinks"]["thumbnail"] rescue "/images/noimage.png"
-  end
-  def value_search(search)
-    search ? search.split('+').join(' ') : nil
-  end
-end
